[MODIFY] request params 의 형태 오류 수정 및 응답 값 검증 수정
