+++ src/test/java/atdd/station/controller/PathAcceptanceTest.java
@@ -1,0 +1,40 @@
+package atdd.station.controller;
+
+import atdd.station.AbstractAcceptanceTest;
+import atdd.station.dto.path.PathFindResponseDto;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.test.web.reactive.server.EntityExchangeResult;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class PathAcceptanceTest extends AbstractAcceptanceTest {
+    private static final String PATH_API_BASE_URL = "/paths";
+
+    private RestWebClientTest restWebClientTest;
+
+    @BeforeEach
+    void setUp() {
+        this.restWebClientTest = new RestWebClientTest(this.webTestClient);
+    }
+
+    @DisplayName("지하철_간의_최단거리를_구한다")
+    @Test
+    void findShortestPath() {
+        long startStationId = 1L;
+        long endStationId = 5L;
+
+        //when
+        EntityExchangeResult<PathFindResponseDto> expectResponse
+                = restWebClientTest.getMethodAcceptance
+                (PATH_API_BASE_URL + "?startId=" + startStationId + "&endId=" + endStationId, PathFindResponseDto.class);
+
+        PathFindResponseDto path = expectResponse.getResponseBody();
+
+        //then
+        assertThat(path.getStartStationId()).isEqualTo(1L);
+        assertThat(path.getEndStationId()).isEqualTo(5L);
+        assertThat(path.getStations().size()).isEqualTo(4);
+    }
+}
